refactor: organize inferred type helpers
