+++ src/zod4/is-compatible-type.ts
@@ -1,158 +1,21 @@
 import type { $ZodObject, $ZodType, $ZodTypes, $ZodUnion } from "zod/v4/core";
 import { createCompareFn } from "./create-compare-fn.ts";
+import {
+  getFiniteLiteralValues,
+  isInferredAsNever,
+  schemaAcceptsNull,
+  schemaAcceptsUndefined,
+  schemaAllowsMissingObjectKey,
+} from "./inferred-type.ts";
 import { isSameType } from "./is-same-type.ts";
 import type { CompareRule } from "./types.ts";
 import {
+  flatUnwrapIntersection,
   flatUnwrapUnion,
-  isInferredAsNever,
+  getInnerType,
   isZodType,
   isZodTypes,
 } from "./utils.ts";
-
-const getInnerType = (schema: $ZodTypes): $ZodTypes | undefined => {
-  const def = schema._zod.def;
-  if (
-    "innerType" in def &&
-    typeof def.innerType === "object" &&
-    isZodType(def.innerType) &&
-    isZodTypes(def.innerType)
-  ) {
-    return def.innerType;
-  }
-  return undefined;
-};
-
-const schemaAcceptsUndefined = (schema: $ZodType): boolean => {
-  if (!isZodType(schema) || !isZodTypes(schema)) return false;
-
-  const def = schema._zod.def;
-  if (
-    def.type === "undefined" ||
-    def.type === "any" ||
-    def.type === "unknown"
-  ) {
-    return true;
-  }
-
-  if (def.type === "literal") {
-    return def.values.includes(undefined);
-  }
-
-  if (def.type === "optional") {
-    return true;
-  }
-
-  if (def.type === "nullable") {
-    const innerType = getInnerType(schema);
-    return innerType ? schemaAcceptsUndefined(innerType) : false;
-  }
-
-  if (def.type === "union") {
-    return flatUnwrapUnion(schema as $ZodUnion).some((option) =>
-      schemaAcceptsUndefined(option),
-    );
-  }
-
-  return false;
-};
-
-const schemaAcceptsNull = (schema: $ZodType): boolean => {
-  if (!isZodType(schema) || !isZodTypes(schema)) return false;
-
-  const def = schema._zod.def;
-  if (def.type === "null" || def.type === "any" || def.type === "unknown") {
-    return true;
-  }
-
-  if (def.type === "literal") {
-    return def.values.includes(null);
-  }
-
-  if (def.type === "optional") {
-    const innerType = getInnerType(schema);
-    return innerType ? schemaAcceptsNull(innerType) : false;
-  }
-
-  if (def.type === "nullable") {
-    return true;
-  }
-
-  if (def.type === "union") {
-    return flatUnwrapUnion(schema as $ZodUnion).some((option) =>
-      schemaAcceptsNull(option),
-    );
-  }
-
-  return false;
-};
-
-/**
- * Checks whether an object property may be omitted for the inferred TypeScript
- * object type.
- *
- * This is narrower than accepting `undefined` as a value. For example,
- * `z.union([z.string(), z.undefined()])` accepts `undefined`, but still infers a
- * required property when used in `z.object({ key: ... })`.
- */
-const schemaAllowsMissingObjectKey = (schema: $ZodType): boolean => {
-  if (!isZodType(schema) || !isZodTypes(schema)) return false;
-
-  const def = schema._zod.def;
-  if (def.type === "optional") {
-    return true;
-  }
-
-  if (def.type === "nullable") {
-    const innerType = getInnerType(schema);
-    return innerType ? schemaAllowsMissingObjectKey(innerType) : false;
-  }
-
-  return false;
-};
-
-const getEnumValues = (
-  entries: Record<string, unknown>,
-): readonly unknown[] => {
-  const values = Object.entries(entries)
-    .filter(([key, value]) => {
-      if (typeof value !== "string") return true;
-      const numericKey = Number(key);
-      return !(
-        Number.isFinite(numericKey) && Object.is(entries[value], numericKey)
-      );
-    })
-    .map(([, value]) => value);
-  return Array.from(new Set(values));
-};
-
-/**
- * Extracts the finite value set represented by literal and enum schemas.
- *
- * These values can be compared with subset logic before falling back to broader
- * kind rules.
- */
-const getFiniteLiteralValues = (
-  schema: $ZodTypes,
-): readonly unknown[] | undefined => {
-  const def = schema._zod.def;
-  if (def.type === "literal") {
-    return def.values;
-  }
-  if (def.type === "enum") {
-    return getEnumValues(def.entries);
-  }
-  return undefined;
-};
-
-const flatUnwrapIntersection = (schema: $ZodTypes): $ZodTypes[] => {
-  const def = schema._zod.def;
-  if (def.type !== "intersection") return [schema];
-  if (!isZodTypes(def.left) || !isZodTypes(def.right)) return [];
-  return [
-    ...flatUnwrapIntersection(def.left),
-    ...flatUnwrapIntersection(def.right),
-  ];
-};
 
 const getReadonlyObjectType = (schema: $ZodTypes): $ZodObject | undefined => {
   const def = schema._zod.def;
