+++ src/zod4/is-same-type.ts
@@ -1,10 +1,10 @@
 import type { $ZodTypes, $ZodUnion } from "zod/v4/core";
 import { isLegacyZodFunction } from "./compat.ts";
 import { createCompareFn } from "./create-compare-fn.ts";
+import { isInferredAsNever } from "./inferred-type.ts";
 import type { CompareRule } from "./types.ts";
 import {
   flatUnwrapUnion,
-  isInferredAsNever,
   isSimpleType,
   isZodType,
   isZodTypes,
